County irrigation attributes

A dataset with the amount of acreage irrigated each year in each county by each category of irrigation distribution system.

Usage

irr_attributes

Format

A data frame with 2,392 rows and 6 variables:

County

The name of the County, in title case ("McCormick")

Category

Irrigation Power Unit Types, Irrigation Wells, or Pumps

Subcategory

Different responses for each attribute

Year

1997-2000

Value

The total or percent response from each county

Source

name of the extension agent who collected the data

Source

W Bryan Smith, Area Extension Agent, Clemson University

Details

Data collected by county extension agents in South Carolina.

References

Irrigation Journal (2001)

See also